moduling

+++ delete.php
@@ -1,10 +1,15 @@
 <?php
+    include_once "db.php"; 
     $id=$_GET['id'];
+    $query="select * from post WHERE id='$id'";
+    $postID=mq($query);
+    $row=mysqli_fetch_array($postID);
 
-    $con=mysqli_connect('localhost','bin','1234','blog');    
-    $sql="DELETE FROM `post` WHERE id=$id";
+    if($_SESSION['userID']===$row['postID']){
+        $sql="DELETE FROM `post` WHERE id=$id";
+    }
 
-    $result =mysqli_query($con,$sql);
+    $result =mq($sql);
     
     if($result){
         echo "삭제 되었습니다.";
@@ -13,6 +18,5 @@
         echo "false";
     }
     echo "<a href=index.php>첫 화면으로 돌아가기</a>";
-
 
 ?>
